refactor(carton): move entry linter scopes into model::entries
